change file names to match convention
